- Moved all image check and build operations to a single struct name `ImageBuilder`. This struct provides just a single `Build` method.
- Added and cleaned tests.
