+++ vm/vm_test.go
@@ -1,20 +1,25 @@
 package vm
 
 import (
-	"fmt"
 	"path/filepath"
 
 	"github.com/ellcrys/druid/util/logger"
 	"github.com/mitchellh/go-homedir"
 	. "github.com/onsi/ginkgo"
-	// . "github.com/onsi/gomega"
+	. "github.com/onsi/gomega"
 )
 
 var _ = Describe("Vm", func() {
-
 	hdir, _ := homedir.Dir()
-
 	log := logger.NewLogrus()
-	vm := New(log, filepath.Join(hdir, "mountdir"))
-	fmt.Println(vm.Init())
+	var vm *VM
+
+	BeforeEach(func() {
+		vm = New(log, filepath.Join(hdir, "mountdir"))
+	})
+
+	BeforeEach(func() {
+		err := vm.Init()
+		Expect(err).To(BeNil())
+	})
 })
